refactor(models): share model selection normalizer
